Merge ILSA data

Merges 'SPSS' data from different International Large-Scale Assessments (ILSA). This function has been tested to behave correctly for: 'TIMSS', 'TIMSS Advanced', 'PIRLS', 'ICCS', 'ICILS', 'CIVED', 'REDS', 'RLII', and 'SITES' (2006).

Usage

ILSAmerge(
  inputdir,
  outputdir,
  population = NULL,
  filetype = c("rds", "zsav", "sav"),
  MBlimit = NULL,
  MBlistlimit = 200,
  SPSSlimit = 50,
  quiet = FALSE
)

Arguments

inputdir

a string indicating the path were ILSA 'SPSS' files are stored.

outputdir

the directory where the merged data will be saved.

population

a character vector indicating which files should be merged. If NULL (the default), all files will be merged. For more information on available populations, run ILSAfile.info() first.

filetype

a string indicating the type of file to be saved, it can be "rds", "zsav", or "sav".

MBlimit

a numerical value indicating the allowed limit of the combined storage of the files of one type (see ILSAfile.info()). For type files that go over the limit, files will not be merged in R, but an 'SPSS' syntax will be produced via spss.syntax(). If set to NULL, no limit will be used and all files will be merged within R. If speed is a problem, we recommend that this number should not be over 200 and merge the rest in 'SPSS'.

MBlistlimit

a numerical value indicating the allowed limit of the combined storage of the files of one type for merging through a list. Values over the limit will be merged through a matrix, which will be slower but uses less memory. Default is 200.

SPSSlimit

a numerical value indicating the limit of files per command of 'SPSS', typically 50.

quiet

a logical value indicating if status of progress should be shown. Default is FALSE.

Value

Saves merged ILSA data or .sps syntax for merging ILSA data.

Examples

# For example, after downloading 'RLII' 1991 G4 data:

# Downloading 'RLII' 1991 and unzipping files
ILSAdownload(study = "RLII", year = 1991, outputdir = tempdir(), unzip = TRUE, agreeLicense = TRUE)
#> By accessing the Data Repository, IDB Analyzer and Data visualizer, you indicate that you agree to the terms and conditions associated with their use. Please read the Disclaimer and License Agreement for full details.
#> 1 files found for RLII 1991.
#> Visit https://www.iea.nl/data-tools/repository/readingliteracy to know how to use and cite these datasets.

# Path were raw 'SPSS' files are
input <- file.path(tempdir(),"RLII1991_IDB_SPSS/Data")

# Path were merged files will be saved
output <- file.path(tempdir(),"RLII1991_IDB_SPSS")

# Merging 'RLII' 1991, as .rds file
ILSAmerge(inputdir = input, outputdir = output, filetype = "rds", quiet = FALSE)
#> 9 files detected. Merging into 1 files.
#> Merging ASCt1. Type 1 of 1.
#> Merging dataset 1 of 9.
#> Merging dataset 2 of 9.
#> Merging dataset 3 of 9.
#> Merging dataset 4 of 9.
#> Merging dataset 5 of 9.
#> Merging dataset 6 of 9.
#> Merging dataset 7 of 9.
#> Merging dataset 8 of 9.
#> Merging dataset 9 of 9.
#> Merging ASCt1 took 3 seconds or 0.05 minutes.
#> Merging took 3 seconds or 0.05 minutes.