'SPSS' merge syntax

Produces and saves an 'SPSS' merge syntax given a list of files.

Usage

spss.syntax(filelist, name, outputdir, zsav = TRUE, SPSSlimit = 50)

Arguments

filelist

a character vector with the list of files to be merged.

name

a string with the name of the merged file (without any extension).

outputdir

the directory where the .sps file and the merged file will be saved.

zsav

a logical value indicating if the the merged file should be compressed with zsav. Default is TRUE.

SPSSlimit

a numerical value indicating the limit of files per command of 'SPSS', typically 50.

Value

Saves an .sps file with the 'SPSS' syntax for merging the desired files.

Examples

# For example, after downloading 'RLII' 1991 G4 data:

# Downloading 'RLII' 1991 and unzipping files
ILSAdownload(study = "RLII", year = 1991, outputdir = tempdir(), unzip = TRUE, agreeLicense = TRUE)
#> By accessing the Data Repository, IDB Analyzer and Data visualizer, you indicate that you agree to the terms and conditions associated with their use. Please read the Disclaimer and License Agreement for full details.
#> 1 files found for RLII 1991.
#> Visit https://www.iea.nl/data-tools/repository/readingliteracy to know how to use and cite these datasets.

# Path were raw 'SPSS' files are
input <- file.path(tempdir(),"RLII1991_IDB_SPSS/Data")

# Path were merged files will be saved
output <- tempdir()

# List of ASCt1 files to be merged
files <- list.files(path = input, pattern = "ASC.+t1")

# Create 'SPSS' syntax
spss.syntax(filelist = files, name = "ASCt1", outputdir = output, zsav = TRUE)